# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Luxury company platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface UserInput {
  email: string;
  name: string;
  /** @minLength 6 */
  password: string;
  role: string;
}

export interface UserUpdate {
  email?: string;
  name?: string;
  role?: string;
}

export interface Project {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  imageUrls?: string | null;
  status: string;
  featured: boolean;
  sortOrder: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ProjectInput {
  title: string;
  slug: string;
  description: string;
  content?: string;
  client?: string;
  year?: number;
  category?: string;
  tags?: string;
  coverImageUrl?: string;
  imageUrls?: string;
  status: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface ProjectUpdate {
  title?: string;
  slug?: string;
  description?: string;
  content?: string;
  client?: string;
  year?: number;
  category?: string;
  tags?: string;
  coverImageUrl?: string;
  imageUrls?: string;
  status?: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface CaseStudy {
  id: number;
  title: string;
  slug: string;
  summary: string;
  challenge: string;
  solution: string;
  outcome: string;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  metrics?: string | null;
  featured: boolean;
  sortOrder: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface CaseStudyInput {
  title: string;
  slug: string;
  summary: string;
  challenge: string;
  solution: string;
  outcome: string;
  client?: string;
  industry?: string;
  duration?: string;
  coverImageUrl?: string;
  tags?: string;
  metrics?: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface CaseStudyUpdate {
  title?: string;
  slug?: string;
  summary?: string;
  challenge?: string;
  solution?: string;
  outcome?: string;
  client?: string;
  industry?: string;
  duration?: string;
  coverImageUrl?: string;
  tags?: string;
  metrics?: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface Service {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  iconName?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  features?: string | null;
  sortOrder: number;
  active: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ServiceInput {
  title: string;
  slug: string;
  description: string;
  content?: string;
  iconName?: string;
  coverImageUrl?: string;
  features?: string;
  sortOrder?: number;
  active?: boolean;
}

export interface ServiceUpdate {
  title?: string;
  slug?: string;
  description?: string;
  content?: string;
  iconName?: string;
  coverImageUrl?: string;
  features?: string;
  sortOrder?: number;
  active?: boolean;
}

export interface Company {
  id: number;
  name: string;
  tagline: string;
  description: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  foundedYear?: number | null;
  /** @nullable */
  teamSize?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  missionStatement?: string | null;
  /** @nullable */
  visionStatement?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface CompanyUpdate {
  name?: string;
  tagline?: string;
  description?: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  country?: string;
  website?: string;
  logoUrl?: string;
  foundedYear?: number;
  teamSize?: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  instagramUrl?: string;
  missionStatement?: string;
  visionStatement?: string;
}

export interface Media {
  id: number;
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  url: string;
  /** @nullable */
  alt?: string | null;
  /** @nullable */
  caption?: string | null;
  createdAt: string;
}

export interface MediaUpdate {
  alt?: string;
  caption?: string;
}

export interface Setting {
  id: number;
  key: string;
  value: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface SettingInput {
  value: string;
  description?: string;
}

export interface DashboardStats {
  totalProjects: number;
  totalCaseStudies: number;
  totalServices: number;
  totalMedia: number;
  featuredProjects: number;
  featuredCaseStudies: number;
  activeServices: number;
}

export type ListProjectsParams = {
featured?: boolean;
status?: string;
};

export type ListCaseStudiesParams = {
featured?: boolean;
};

export type ListMediaParams = {
type?: string;
};